Prevent potential invalid pointer dereference.
This change fixes a potential issue where the returned event was the address of
an `eventstore.Event` within a slice, which may be invalidated when more events
are appended.

I've changed the internal representation of the event-store to keep a slice of
pointers. This should resist the introduction of similar bugs, which came about
because the public API always uses pointers to events.

Finally, the event is cloned on the way out of `Result.Next()` to ensure that
any modifications do not affect the "persisted" event.