Add error for not finding a message by ID.
Previously all implementations of
`eventstore.Repository.LoadEventBySource()` method did not return any
error if the barrier message was not found. With the addition of
`sql.Driver.SelectOffsetByMessageID()` it became necessary to check if the
barrier message exists at all.

This commit adds the return of the error in `boltdb` and `memory`
providers to be consistent with SQL providers.

It also adds a test to check if the error is returned.